forgot to remove all the files in the control directory
If the lock file is downloaded and the mcmc is locked, we need to
remove the lock file before we try to do the mcmc again. Otherwise, the
mcmc will end up in a neverending locked loop because scp only copies
down the actual files, it doesn't look to see if an existing file on
the local machine actually exists on the remote machine.